+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/UserImpl.kt
@@ -16,12 +16,16 @@
 
 package io.realm.kotlin.mongodb.internal
 
+import io.realm.kotlin.internal.interop.sync.AuthProvider
 import io.realm.kotlin.internal.interop.RealmInterop
 import io.realm.kotlin.internal.interop.RealmUserPointer
 import io.realm.kotlin.internal.interop.sync.CoreUserState
 import io.realm.kotlin.internal.platform.freeze
 import io.realm.kotlin.internal.util.use
+import io.realm.kotlin.mongodb.AuthenticationProvider
+import io.realm.kotlin.mongodb.Credentials
 import io.realm.kotlin.mongodb.User
+import io.realm.kotlin.mongodb.UserIdentity
 import kotlinx.coroutines.channels.Channel
 
 // TODO Public due to being a transitive dependency to SyncConfigurationImpl
@@ -35,10 +39,28 @@ public class UserImpl(
 
     // TODO Can maybe fail, but we could also cache the return value?
     override val identity: String
+        get() = id
+    override val id: String
         get() = RealmInterop.realm_user_get_identity(nativePointer)
-
     override val loggedIn: Boolean
         get() = RealmInterop.realm_user_is_logged_in(nativePointer)
+    override val identities: List<UserIdentity>
+        get() = RealmInterop.realm_user_get_all_identities(nativePointer).map {
+            val authProvider = when(it.provider) {
+                AuthProvider.RLM_AUTH_PROVIDER_ANONYMOUS -> AuthenticationProvider.ANONYMOUS
+                AuthProvider.RLM_AUTH_PROVIDER_ANONYMOUS_NO_REUSE -> AuthenticationProvider.ANONYMOUS
+                AuthProvider.RLM_AUTH_PROVIDER_FACEBOOK -> AuthenticationProvider.FACEBOOK
+                AuthProvider.RLM_AUTH_PROVIDER_GOOGLE -> AuthenticationProvider.GOOGLE
+                AuthProvider.RLM_AUTH_PROVIDER_APPLE -> AuthenticationProvider.APPLE
+                AuthProvider.RLM_AUTH_PROVIDER_CUSTOM -> TODO()
+                AuthProvider.RLM_AUTH_PROVIDER_EMAIL_PASSWORD -> AuthenticationProvider.EMAIL_PASSWORD
+                AuthProvider.RLM_AUTH_PROVIDER_FUNCTION -> TODO()
+                AuthProvider.RLM_AUTH_PROVIDER_USER_API_KEY -> AuthenticationProvider.API_KEY
+                AuthProvider.RLM_AUTH_PROVIDER_SERVER_API_KEY -> TODO()
+                else -> throw IllegalStateException("Unknown auth provider: ${it.provider}")
+            }
+            UserIdentity(it.id, authProvider)
+        }
 
     override suspend fun logOut() {
         Channel<Result<Unit>>(1).use { channel ->
@@ -84,6 +106,25 @@ public class UserImpl(
             return@use channel.receive()
                 .getOrThrow()
         }
+    }
+
+    override suspend fun linkCredentials(credentials: Credentials): User {
+        if (state != User.State.LOGGED_IN) {
+            throw IllegalStateException("User must be logged in, in order to link credentials to it.")
+        }
+        Channel<Result<Unit>>(1).use { channel ->
+            RealmInterop.realm_app_link_credentials(
+                app.nativePointer,
+                nativePointer,
+                (credentials as CredentialsImpl).nativePointer,
+                channelResultCallback<Unit, Unit>(channel) {
+                    // No-op
+                }.freeze()
+            )
+            return@use channel.receive()
+                .getOrThrow()
+        }
+        return this
     }
 
     override fun equals(other: Any?): Boolean {